backend/requirements.txt
fastapi[standard]
pydantic-settings
uvicorn
requests
python-dotenv
